Add error checking for an invalid lookup key
If the lookup key specified within a component is falsy (such as being an empty string, null, undefined, or a few other conditions), i18n.js will output an error in the page's console.

This will not stop execution of the script, which will instead try to use the fallback string or otherwise output "UNDEFINED".