Remove separator configuration option
Removes the config option to change the separators used by placeholder strings. This is due to placeholders also being present in hardcoded fallback strings, which couldn't easily be changed if
the separator configuration was changed.

As customising separators was already only a nice to have, I've removed the configuration option for now.